Split FinOpsContent.xaml.cs into per-concern partials
Mechanical, behavior-neutral partial-class split of the FinOpsContent WPF
code-behind. No logic or call-site changes — methods (with their preceding
comments) moved verbatim into sibling partial files of the same class:

- FinOpsContent.Loaders.cs    — data-loading methods + helpers
- FinOpsContent.Refresh.cs    — refresh button + time-range change handlers
- FinOpsContent.CopyExport.cs — copy/export/clipboard context-menu handlers
- FinOpsContent.Filters.cs    — column-filtering region + DbSize filter handlers

The main FinOpsContent.xaml.cs retains only fields, ctor, OnLoaded,
Initialize, ServerSelector_SelectionChanged, and RefreshDataAsync
(1268 -> 148 lines).

Verified: dotnet build PerformanceMonitor.sln -c Debug -> Build succeeded,
0 warnings, 0 errors. dotnet test Dashboard.Tests -> 488 passed, 0 failed.
