+++ Dashboard/Controls/FinOpsContent.CopyExport.cs
@@ -1,0 +1,173 @@
+/*
+ * Copyright (c) 2026 Erik Darling, Darling Data LLC
+ *
+ * This file is part of the SQL Server Performance Monitor.
+ *
+ * Licensed under the MIT License. See LICENSE file in the project root for full license information.
+ */
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Controls.Primitives;
+using System.Windows.Data;
+using System.Windows.Media;
+using Microsoft.Win32;
+using PerformanceMonitorDashboard.Helpers;
+using PerformanceMonitorDashboard.Models;
+using PerformanceMonitorDashboard.Services;
+using PerformanceMonitor.Ui;
+using PerformanceMonitor.Common;
+
+namespace PerformanceMonitorDashboard.Controls
+{
+    public partial class FinOpsContent : UserControl
+    {
+
+        // ============================================
+        // Copy / Export Context Menu Handlers
+        // ============================================
+
+        private static DataGrid? FindParentDataGrid(DependencyObject? element)
+        {
+            while (element != null)
+            {
+                if (element is DataGrid dg) return dg;
+                element = VisualTreeHelper.GetParent(element);
+            }
+            return null;
+        }
+
+        private void CopyCell_Click(object sender, RoutedEventArgs e)
+        {
+            if (sender is MenuItem menuItem && menuItem.Parent is ContextMenu contextMenu)
+            {
+                var grid = FindParentDataGrid(contextMenu.PlacementTarget);
+                if (grid != null && grid.CurrentCell.Column != null)
+                {
+                    var cellContent = TabHelpers.GetCellContent(grid, grid.CurrentCell);
+                    if (!string.IsNullOrEmpty(cellContent))
+                    {
+                        /* Use SetDataObject with copy=false to avoid WPF's problematic Clipboard.Flush() */
+                        Clipboard.SetDataObject(cellContent, false);
+                    }
+                }
+            }
+        }
+
+        private void CopyRow_Click(object sender, RoutedEventArgs e)
+        {
+            if (sender is MenuItem menuItem && menuItem.Parent is ContextMenu contextMenu)
+            {
+                var grid = FindParentDataGrid(contextMenu.PlacementTarget);
+                if (grid != null && grid.SelectedItem != null)
+                {
+                    var rowText = TabHelpers.GetRowAsText(grid, grid.SelectedItem);
+                    if (!string.IsNullOrEmpty(rowText))
+                    {
+                        /* Use SetDataObject with copy=false to avoid WPF's problematic Clipboard.Flush() */
+                        Clipboard.SetDataObject(rowText, false);
+                    }
+                }
+            }
+        }
+
+        private void CopyAllRows_Click(object sender, RoutedEventArgs e)
+        {
+            if (sender is MenuItem menuItem && menuItem.Parent is ContextMenu contextMenu)
+            {
+                var grid = FindParentDataGrid(contextMenu.PlacementTarget);
+                if (grid != null)
+                {
+                    var sb = new StringBuilder();
+
+                    // Header row
+                    var headers = grid.Columns.Select(c => DataGridClipboardBehavior.GetHeaderText(c));
+                    sb.AppendLine(string.Join("\t", headers));
+
+                    // Data rows
+                    foreach (var item in grid.Items)
+                    {
+                        var values = new List<string>();
+                        foreach (var column in grid.Columns)
+                        {
+                            var binding = (column as DataGridBoundColumn)?.Binding as Binding;
+                            if (binding != null)
+                            {
+                                var prop = item.GetType().GetProperty(binding.Path.Path);
+                                var value = prop?.GetValue(item)?.ToString() ?? string.Empty;
+                                values.Add(value);
+                            }
+                        }
+                        sb.AppendLine(string.Join("\t", values));
+                    }
+
+                    if (sb.Length > 0)
+                    {
+                        /* Use SetDataObject with copy=false to avoid WPF's problematic Clipboard.Flush() */
+                        Clipboard.SetDataObject(sb.ToString(), false);
+                    }
+                }
+            }
+        }
+
+        private void ExportToCsv_Click(object sender, RoutedEventArgs e)
+        {
+            if (sender is MenuItem menuItem && menuItem.Parent is ContextMenu contextMenu)
+            {
+                var grid = FindParentDataGrid(contextMenu.PlacementTarget);
+                if (grid != null)
+                {
+                    var dialog = new SaveFileDialog
+                    {
+                        Filter = "CSV files (*.csv)|*.csv|All files (*.*)|*.*",
+                        DefaultExt = ".csv",
+                        FileName = $"FinOps_Export_{DateTime.Now:yyyyMMdd_HHmmss}.csv"
+                    };
+
+                    if (dialog.ShowDialog() == true)
+                    {
+                        try
+                        {
+                            var sb = new StringBuilder();
+
+                            // Header row
+                            var sep = TabHelpers.CsvSeparator;
+                            var headers = grid.Columns.Select(c => TabHelpers.EscapeCsvField(DataGridClipboardBehavior.GetHeaderText(c), sep));
+                            sb.AppendLine(string.Join(sep, headers));
+
+                            // Data rows
+                            foreach (var item in grid.Items)
+                            {
+                                var values = new List<string>();
+                                foreach (var column in grid.Columns)
+                                {
+                                    var binding = (column as DataGridBoundColumn)?.Binding as Binding;
+                                    if (binding != null)
+                                    {
+                                        var prop = item.GetType().GetProperty(binding.Path.Path);
+                                        values.Add(TabHelpers.EscapeCsvField(TabHelpers.FormatForExport(prop?.GetValue(item)), sep));
+                                    }
+                                }
+                                sb.AppendLine(string.Join(sep, values));
+                            }
+
+                            File.WriteAllText(dialog.FileName, sb.ToString());
+                        }
+                        catch (Exception ex)
+                        {
+                            Logger.Error($"Error exporting to CSV: {ex.Message}", ex);
+                            MessageBox.Show($"Error exporting to CSV: {ex.Message}", "Export Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                        }
+                    }
+                }
+            }
+        }
+
+    }
+}
